Separate schemas by endpoint
