+++ src/index.ts
@@ -1,3 +1,6 @@
 export { OpenAIClient } from './openai-client';
-export { CompletionParams, EmbeddingParams } from './schema';
-export type { EmbeddingResponse, CompletionResponse } from './schema';
+export { CompletionParams } from './schemas/completion';
+export { EmbeddingParams } from './schemas/embedding';
+
+export type { CompletionResponse } from './schemas/completion';
+export type { EmbeddingResponse } from './schemas/embedding';
